arrR: Individual-Based Simulation Model of Artifical Reefs in R

Individual-based simulation model to analyze how artificial reefs in combination with a fish population influence seagrass growth.

For more information about how to use the model, please see https://allgeier-lab.github.io/arrR/.

See also

Author

Maintainer: Maximilian H.K. Hesselbarth mhk.hesselbarth@gmail.com (ORCID)

Authors:

Kenzo E. Esquivel kesquivel@berkeley.edu (ORCID)

Katrina S. Munsterman kmunster@umich.edu (ORCID)

Other contributors:

Samantha Iliff siliff@umich.edu (ORCID) [illustrator]

Jacob E. Allgeier jeallg@umich.edu (ORCID) [research team head]