check_parameters

Check model parameters and starting values.

Usage

check_parameters(starting_values = NULL, parameters = NULL, verbose = TRUE)

Arguments

starting_values

List with all starting value parameters.

parameters

List with all model parameters.

verbose

Logical if TRUE, progress reports are printed.

Value

void

Details

The function checks if all required starting values and parameters are provided. If parameters and/or starting values are missing, a vector with the name of the corresponding values is returned.

If no arguments are provided, a list of all required parameters and values is printed.

Examples

check_parameters()
#> > Required starting values:
#> bg_biomass
#> ag_biomass
#> nutrients_pool
#> detritus_pool
#> pop_n
#> pop_mean_size
#> pop_sd_size
#> 
#> > Required parameters:
#> bg_biomass_min
#> bg_biomass_max
#> bg_v_max
#> bg_k_m
#> bg_gamma
#> ag_biomass_min
#> ag_biomass_max
#> ag_v_max
#> ag_k_m
#> ag_gamma
#> seagrass_thres
#> seagrass_slope
#> seagrass_slough
#> nutrients_diffusion
#> nutrients_loss
#> detritus_mineralization
#> detritus_diffusion
#> detritus_fish_decomp
#> detritus_fish_diffusion
#> detritus_loss
#> move_mean
#> move_sd
#> move_border
#> move_reef
#> move_return
#> pop_reserves_max
#> pop_reserves_thres_mean
#> pop_reserves_thres_sd
#> pop_reserves_consump
#> pop_a
#> pop_b
#> pop_k
#> pop_linf
#> pop_n_body
#> resp_intercept
#> resp_slope
#> resp_temp_low
#> resp_temp_optm
#> resp_temp_max