Default starting values

List with default starting values.

Usage

default_starting

Format

List

References

Citations for all parameters and starting values can be found in Esquivel, K.E., Hesselbarth, M.H.K., Allgeier, J.E., 2022. Mechanistic support for increased primary production around artificial reefs. Ecological Applications e2617. <https://doi.org/10.1002/eap.2617>