get_starting_values

Get mean starting values of model run.

Usage

get_starting_values(seafloor_values, fishpop_values)

Arguments

seafloor_values

Matrix with seafloor values.

fishpop_values

Matrix population values.

Value

list

Details

Get mean starting values. Includes values for i) bg_biomass, ii) ag_biomass, iii) nutrients_pool, iv) detritus_pool, v) pop_n, and vi) pop_mean_size. The function is needed because some of the values can include stochasticity and can differ from the provided starting values.

Examples

if (FALSE) { # \dontrun{
get_starting_values(seafloor_values = seafloor_values, fishpop_values = fishpop_values)
} # }