rcpp_move_behav

Rcpp simulate movement (behav).

Usage

rcpp_move_behav(
  fishpop,
  fishpop_attr,
  move_mean,
  move_sd,
  move_reef,
  move_border,
  move_return,
  max_dist,
  coords_reef,
  extent,
  dimensions
)

Arguments

fishpop

Matrix with fishpop values.

fishpop_attr

Matrix with id, pop_reserves_thres_mean, and pop_reserves_consump values

move_mean, move_sd

Double with mean movement parameter.

move_reef

Double with mean movement distance when sheltering at reef.

move_border

Double with movement distance that surrounds reef cell border.

move_return

Double with mean movement distance when returning to reef.

max_dist

Maximum distance an individual can move.

coords_reef

Matrix with ID and coords of reef cells.

extent

Vector with extent (xmin,xmax,ymin,ymax).

dimensions

Vector with dimensions (nrow, ncol).

Value

void

Details

Fish individuals move based on how much nutrients they have stored in their reserves. There are three different movement states.

If reserves are above a certain threshold, individuals either shelter at reef cells (state 1) or move back towards reef cells (state 2). If reserves are not above the threshold, individuals move randomly across the seafloor to forage (state 3).