rcpp_simulate

Rcpp run simulation.

Usage

rcpp_simulate(
  seafloor,
  fishpop,
  nutrients_input,
  seafloor_track,
  fishpop_track,
  parameters,
  fishpop_attr,
  movement,
  extent,
  dimensions,
  max_i,
  min_per_i,
  save_each,
  seagrass_each,
  burn_in,
  to_disk,
  path_disk,
  verbose
)

Arguments

seafloor, fishpop

Matrix with seafloor and fishpop data.

nutrients_input

Vector with amount of nutrient input each time step.

seafloor_track, fishpop_track

List with entry for each saving time step.

parameters

List with parameters.

fishpop_attr

Matrix with id, pop_reserves_thres_mean, and pop_reserves_consump values

movement

String specifing movement algorithm.

extent

Vector with extent (xmin,xmax,ymin,ymax).

dimensions

Vector with dimensions (nrow, ncol).

max_i

Integer with maximum number of simulation time steps.

min_per_i

Integer to specify minutes per i.

save_each

Numeric how often data should be saved to return.

seagrass_each

Integer how often (each i * x) seagrass dynamics will be simulated.

burn_in

Numeric with time steps used to burn in.

to_disk

Logical if TRUE, results are written into a text file.

verbose

Logical if TRUE, progress reports are printed.

disk_path

String with path to result text files.

Value

void

Details

The functions is a 'wrapper' around the following sub-processes: (i) nutrient input, (ii) seagrass growth, (iii) detritus mineralization, (iv) movement of individuals, (v) respiration of individuals, (vi) growth of individuals, (vii) mortality of individuals, (viii) diffusion of nutrients/detritus, and ix) nutrient output.

References

For a detailed model description, please see Esquivel, K.E., Hesselbarth, M.H.K., Allgeier, J.E., 2022. Mechanistic support for increased primary production around artificial reefs. Ecological Applications e2617. <https://doi.org/10.1002/eap.2617>